refactor(ui): refactor Avatar
