+++ ui/design/src/components/Avatar/__test__/index.test.tsx
@@ -1,12 +1,10 @@
 import '@testing-library/jest-dom/extend-expect';
-import { cleanup, fireEvent, waitForElement } from '@testing-library/react';
+import { cleanup, waitForElement } from '@testing-library/react';
 import * as React from 'react';
 import { act, create, ReactTestRenderer } from 'react-test-renderer';
-import { createFile, customRender, delay, wrapWithTheme } from '../../../test-utils';
-import { MockFileReader, WindowWithFileReader } from '../../../test-utils/mocks';
-import Avatar, { getAvatarFromSeed } from '../avatar';
+import { customRender, delay, wrapWithTheme } from '../../../test-utils';
+import Avatar, { getAvatarFromSeed } from '..';
 import AvatarImage from '../avatar-image';
-import EditableAvatar from '../editable-avatar';
 
 describe('<Avatar /> component', () => {
   let componentWrapper: ReactTestRenderer = create(<></>);
@@ -61,52 +59,5 @@ describe('<Avatar /> component', () => {
     );
     const image = await waitForElement(() => findByTestId('avatar-image'));
     expect(image.getAttribute('src')).toEqual(src);
-  });
-});
-
-describe('<EditableAvatar /> Component', () => {
-  let componentWrapper: ReactTestRenderer = create(<></>);
-  const originalFileReader = FileReader;
-
-  beforeEach(() => {
-    (window as WindowWithFileReader).FileReader = MockFileReader;
-    act(() => {
-      componentWrapper = create(
-        wrapWithTheme(
-          <EditableAvatar onChange={jest.fn()} ethAddress={'0x1230am3421h3i14cvv21n4'} />,
-        ),
-      );
-    });
-  });
-  afterEach(() => {
-    (window as WindowWithFileReader).FileReader = originalFileReader;
-    act(() => {
-      componentWrapper.unmount();
-    });
-    cleanup();
-  });
-  it('should match snapshot', async () => {
-    // delay to allow avatar to mount the AvatarImage component
-    await delay();
-    expect(componentWrapper.toJSON()).toMatchSnapshot('editable-avatar');
-  });
-  it('should have 1 input type file', async () => {
-    const { getAllByTestId } = customRender(
-      <EditableAvatar onChange={jest.fn()} ethAddress={'0x1230am3421h3i14cvv21n4'} />,
-      {},
-    );
-    const fileInput = await waitForElement(() => getAllByTestId('avatar-file-input'));
-    expect(fileInput).toHaveLength(1);
-    expect(fileInput[0].getAttribute('type')).toEqual('file');
-  });
-  it('should trigger onChange event when input is changed', async () => {
-    const onChange = jest.fn();
-    const { findByTestId } = customRender(
-      <EditableAvatar onChange={onChange} ethAddress={'0x1230am3421h3i14cvv21n4'} />,
-      {},
-    );
-    const fileInput = await waitForElement(() => findByTestId('avatar-file-input'));
-    fireEvent.change(fileInput, { target: { file: createFile('test-file.png') } });
-    expect(onChange).toBeCalledTimes(1);
   });
 });
