--- ui/design/src/components/Avatar/index.ts
@@ -1,5 +1,0 @@
-import Avatar from './avatar';
-import EditableAvatar from './editable-avatar';
-import StackedAvatar from './stacked-avatar';
-
-export { Avatar, EditableAvatar, StackedAvatar };
