+++ ui/design/src/components/Avatar/index.tsx
