+++ ui/design/src/components/Buttons/avatar-buttons/profile-avatar-button.tsx
@@ -1,7 +1,7 @@
 import { Box } from 'grommet';
 import * as React from 'react';
 import { truncateMiddle } from '../../../utils/string-utils';
-import { Avatar } from '../../Avatar/index';
+import Avatar from '../../Avatar';
 import { AvatarSize } from '../../Avatar/styled-avatar';
 import StyledIconLink from '../icon-buttons/styled-icon-link';
 import { ButtonInfo, StyledWrapperBox } from './styled-profile-avatar-button';
