+++ ui/design/src/components/Cards/profile-cards/profile-search-card.tsx
@@ -5,7 +5,7 @@ import { IProfileWidgetCard } from './profile-widget-card';
 import { SearchProfileAvatarDiv, StyledInlineBox } from './styled-profile-card';
 
 import { MainAreaCardBox, StyledAnchor } from '../common/basic-card-box';
-import { Avatar } from '../../Avatar/index';
+import Avatar from '../../Avatar';
 import { DuplexButton } from '../../Buttons/index';
 import { Icon } from '../../Icon';
 import { truncateMiddle } from '../../../utils/string-utils';
