+++ ui/design/src/components/Input/comment-input/comment-input.tsx
@@ -1,7 +1,7 @@
 import { Box, Text, TextArea } from 'grommet';
 import React, { useRef, useState } from 'react';
 import useSimpleClickAway from '../../../utils/simpleClickAway';
-import { Avatar } from '../../Avatar/index';
+import Avatar from '../../Avatar';
 import { Icon } from '../../Icon/index';
 import { StyledDiv, StyledCommentWrapper } from './styled-comment-input';
