+++ ui/design/src/components/StackedAvatar/index.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable */
 import * as React from 'react';
-import { Avatar } from './index';
+import Avatar from '../Avatar';
 import { StyledStackBox, StyledContainer } from './styled-stacked-avatar';
 
 export interface IStackedAvatarProps {
