+++ ui/design/src/components/StackedAvatar/styled-stacked-avatar.ts
